refactor(some-censor): adopt commons keybinding/command typestate
some-censor was the workspace the commons keybinding typestate was lifted
from (#276), so it is the lowest-risk first adopter. Replace its local
key-binding.ts with @some-extension/common's typestate: ModifierSet, the
code-based KeyBinding table type, the platform-normalized matcher, and the
input-context guard now come from commons only.

Per Good-Citizen idiom #3, the command handlers stay local: a new
content/commands.ts owns censor's command id (advance-all-to-title) and
binds it to VideoManager.advanceAllToTitle(), then defers to the commons
attachKeyBindings(registry, bindings). The adapter direction is preserved
(commands knows VideoManager; VideoManager knows nothing about keybindings).

The default shortcut is preserved byte-for-byte: Ctrl+Shift+Period
(⌘+Shift+Period on macOS via the commons matcher).

Closes #277
